+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add api endpoint /me/org_topics/ [#3070](https://github.com/opendatateam/udata/pull/3070)
 - Expose dataservices in RDF catalog [#3058](https://github.com/opendatateam/udata/pull/3058) [#3075](https://github.com/opendatateam/udata/pull/3075)
 - CORS: always returns 204 on OPTIONS request [#3046](https://github.com/opendatateam/udata/pull/3046) 
+- Allow discussion notification customization [#3072](https://github.com/opendatateam/udata/pull/3072)
 
 ## 9.0.0 (2024-06-07)